backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn>=0.30.0
gunicorn>=23.0.0
python-multipart>=0.0.10
firebase-admin>=6.5.0
google-genai>=0.2.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.0
pillow>=10.4.0
